LayoutTests/fast/canvas/webgl/canvas-zero-size-expected.txt

Tests that a zero size canvas does not fail.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS program != null is true
PASS getError was expected value: NO_ERROR : Should be no errors from setup.
PASS successfullyParsed is true

TEST COMPLETE
